iteration2
fixed header size, added hrefs, fixed xfr logo,

rob, please try and fix the floating nav at the top for responsive…when
i float the logo left, the nav gets fucked up, but if i put the nav
before the logo in the markup and float it right, it's fine...but
doesn't work for responsive…lame.

also, please work on parallax and carousel.

bw