Added Featured Designers Section, slow scrolling, fixed nav
Added content for featured designers section (made this parallax as
opposed to carosel for now) just because its a parallax site, i felt it
made more sense.

Added slow scroll script so that when any navigation is click on it
slowly scrolls to that element on the page.

Added fixed navigation for the main nav as well as turned the designers
section in to a fixed nav when viewing the featured designers. (I feel
this helped with the parallax illusion as well as it is helpful to the
user)