more changes
- On the ABOUT TOWNHOME Page, they changed "she'd" to "she had" (which
I appreciate!) BUT they put the requested comma in the wrong
spot...they put it after Center City in the paragraph that is on the
Header page (which they now also need to remove) instead of after
"Center City Philadelphia" on the ABOUT TOWNHOME page.

- Some links are still not working correctly - Pinterest, in all 3 of
the link locations; the link on the ABOUT TOWNHOME page,  the icon on
the CONTACT page and also the link on the CONTACT page at the bottom
right. It seems to go to the Dana Bank Pinterest page BUT it's not the
complete page.  Also, the phone icon goes to a blank page.

- Wetalked about swapping placement of "ALL CONTENT COPYRIGHT..." with
"TownHome is 10 years old this year..." at the bottom of the CONTACT
page so it would make more sense.

- Not sure the following header wording is right... "WE'RE CELEBRATING
OUR TENTH ANNNIVERSARY THIS YEAR! CELEBRATE WITH US ON FACEBOOK +
TWITTER + PINTEREST + INSTAGRAM"  I think it sounds repetative. Should
read "JOIN US IN CELEBRATING OUR TENTH ANNIVERSARY ON FACEBOOK +
TWITTER + PINTEREST + INSTAGRAM"

- All the "SEE MORE OF THIS COLLECTION AND OTHERS" should be linked to
TH's home page on FB not the albums and/or featured albums as it does
now...many of the links aren't even going to the right album